refactor(pedagogy): extract prompt_builder tests to sibling file
prompt_builder.rs was 2042 lines, dominated by a ~1390-line
`#[cfg(test)] mod tests` (characterization tests pinning decide_intent's
Socratic-intent heuristic plus prompt-building tests). Move that module
verbatim into a sibling `prompt_builder/tests.rs`, using the idiomatic
Rust-2018 `prompt_builder.rs` + `prompt_builder/tests.rs` form — same
pattern as PRs #268/#269/#270.

prompt_builder.rs drops 2042 -> 652 lines (production code only). The
`#[cfg(test)] fn is_factual_question` helper stays in the parent;
descendant modules reach ancestor items via `use super::*` regardless of
visibility, so no signature changed. Pure test-only relocation: no
behaviour change, no incidental fixes.

Verification: 62 prompt_builder tests pass identically (62 before = 62
after); full workspace `cargo test` 0 failures; clippy + fmt clean.
